move ActivityIndicator directory into Indicator directory
