+++ entities/Indicator/ActivityIndicator/DateRange.ts
